refactor: migrated top level style endpoints
